Rename error.rs to errors.rs
